utils/requirements.txt
aiogram
requests
psutil
aiosqlite
fluent.runtime
g4f[all]